fixup! refactor: ensure fmt string is hardcoded in logger

+++ libtransmission/log.c
@@ -138,7 +138,7 @@ char* tr_logGetTimeStr(char* buf, size_t buflen)
     struct timeval tv;
     tr_gettimeofday(&tv);
     time_t const seconds = tv.tv_sec;
-    int const milliseconds = tv.tv_usec / 1000;
+    int const milliseconds = (int)(tv.tv_usec / 1000);
     char msec_str[8];
     tr_snprintf(msec_str, sizeof msec_str, "%03d", milliseconds);
